Increase robustness of file transfers to remote
If the remote is slow initializing to receive a file, the first
buffer sent from host to remote may overflow the UART buffer.
To fix this, the "ack" mechanism was moved to the beginning of
the loop so the first buffer write waits for the remote.

In debugging this, it was observed that the dev.read() of the ack
would not timeout causing an indefinite "hang". A timeout was
added to send_file_to_remote() to catch issues on the remote.
The timeout methods were changed to getter/setter properties to
enable (re)setting the timeout to None which indicates block
forever on a serial read.