Improve reporting for the "rsync" primitive
When uploading a larger code base to the device, rshell reports well
about "Adding ..." files when actually uploading them. However, if
timestamps haven't changed at all, it would be silent for a while,
while checking all the files about whether they should be updated
without giving appropriate feedback to the user.

This change introduces a "Checking ..." message for each file
in order to be more verbose about what is actually going on.