User's editor as a command
Add a method to call the user's editor by its name.
Example: by default vi is the default editor so "vi filename" call the editor like "edit" would do. If the user override with "--editor nano" then "nano filename" call the editor in the same way.
I did not success to add the EDITOR keyword in the help message.